Add `device_guard` anywhere a `CUDA` function is called.
